Organize file structure.
