+++ AnglePanGestureRecognizer/Extensions/ClosedCountableRange+Grid.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension CountableClosedRange {
+public extension CountableClosedRange {
 
     // swiftlint:disable:next large_tuple
     var grid: [(Bound, Bound, Bound)] {
