+++ AnglePanGestureRecognizer/Model/GridTile.swift
@@ -1,0 +1,21 @@
+//
+//  GridTile.swift
+//  AnglePanGestureRecognizer
+//
+//  Created by Matthew Buckley on 5/5/17.
+//  Copyright © 2017 Atelier Clockwork. All rights reserved.
+//
+
+import Foundation
+
+public struct GridTile<TileData> {
+
+    var point: GridPoint
+    var value: TileData
+
+    init(point: GridPoint, value: TileData) {
+        self.point = point
+        self.value = value
+    }
+
+}
